requirements.txt
discord.py
colorama
nekos.py